optimization: template_base module extraction
