+++ gcloud/iam_auth/view_interceptors/common_template.py
@@ -18,7 +18,7 @@ from iam.exceptions import AuthFailedException, MultiAuthFailedException
 
 from gcloud.utils.strings import string_to_boolean
 from gcloud.common_template.models import CommonTemplate
-from gcloud.common_template.utils import read_template_data_file
+from gcloud.template_base.utils import read_template_data_file
 
 from gcloud.iam_auth import IAMMeta
 from gcloud.iam_auth import res_factory
