+++ gcloud/periodictask/resources.py
@@ -39,7 +39,7 @@ from gcloud.utils.strings import standardize_name
 from gcloud.constants import PERIOD_TASK_NAME_MAX_LENGTH
 from gcloud.core.resources import ProjectResource
 from gcloud.common_template.models import CommonTemplate
-from gcloud.common_template.utils import replace_template_id
+from gcloud.template_base.utils import replace_template_id
 from gcloud.commons.tastypie import GCloudModelResource
 from gcloud.iam_auth import res_factory
 from gcloud.iam_auth import IAMMeta, get_iam_client
